hf-spaces/requirements.txt
# Polymarket Super Bot Requirements

# Core
gradio>=4.0.0
aiohttp>=3.8.0
websockets>=12.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Async
asyncio>=3.4.3

# Logging
loguru>=0.7.0

# Environment
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0

# Type hints
typing-extensions>=4.8.0

# Date/Time
python-dateutil>=2.8.0
